Converted collision to be a component
Now collision is a component (circle_collision) instead of a behaviour
of the interface element. This makes it more customizable as not every
element in the game needs to have collision. Besides that, in the future
there will be more types of collision, such as rectangle, polygon etc...
so the user can choose what type of collision he wants to add to
different elements in the game.